Start of part 2 assignment for week 5.
Link style.css and bootstrap in the head of index.html.
Add header, main and footer to index.html.
Add other elements to header and footer.

Add a few things to style.css to get rid of margin on Body and organize the footer.
Add images to the assignment file of Week 5 Part 2.